# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Dreamshelix India EdTech Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /referrals/my:
    get:
      operationId: getMyReferrals
      tags: [referrals]
      summary: Get my referral details
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Referral details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralDetails"
  /referrals/earnings:
    get:
      operationId: getReferralEarnings
      tags: [referrals]
      summary: Get referral earnings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Referral earnings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EarningsResponse"
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: Get payment history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Payment history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{paymentId}/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify payment
      security:
        - bearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentRequest"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /withdrawals:
    get:
      operationId: listWithdrawals
      tags: [payments]
      summary: Get withdrawal history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
    post:
      operationId: requestWithdrawal
      tags: [payments]
      summary: Request a withdrawal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequest"
      responses:
        "201":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
...
  /admin/withdrawals:
    get:
      operationId: listAdminWithdrawals
      tags: [admin]
      summary: List all withdrawals (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ReferralDetails:
      type: object
      properties:
        referralCode:
          type: string
        referralLink:
          type: string
        totalReferrals:
          type: integer
        directReferrals:
          type: integer
        indirectReferrals:
          type: integer
        totalEarnings:
          type: number
        pendingEarnings:
          type: number
        referredUsers:
          type: array
          items:
            $ref: "#/components/schemas/ReferredUser"
      required:
        - referralCode
        - referralLink
        - totalReferrals
        - directReferrals
        - indirectReferrals
        - totalEarnings
        - pendingEarnings
        - referredUsers
    ReferredUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        joinedAt:
          type: string
          format: date-time
        level:
          type: string
          enum: [direct, indirect]
        commissionEarned:
          type: number
      required:
        - id
        - name
        - joinedAt
        - level
        - commissionEarned
    EarningsResponse:
      type: object
      properties:
        walletBalance:
          type: number
        totalEarnings:
          type: number
        thisMonthEarnings:
          type: number
        pendingEarnings:
          type: number
        earningsHistory:
          type: array
          items:
            $ref: "#/components/schemas/EarningEntry"
      required:
        - walletBalance
        - totalEarnings
        - thisMonthEarnings
        - pendingEarnings
        - earningsHistory
    EarningEntry:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        type:
          type: string
          enum: [direct_commission, indirect_commission]
        status:
          type: string
          enum: [pending, credited]
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - amount
        - type
        - status
        - createdAt
    Payment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, success, failed]
        gateway:
          type: string
        transactionId:
          type: string
        invoiceUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - courseId
        - amount
        - status
        - createdAt
    CreatePaymentRequest:
      type: object
      properties:
        courseId:
          type: integer
        gateway:
          type: string
          enum: [razorpay, stripe]
      required:
        - courseId
        - gateway
    VerifyPaymentRequest:
      type: object
      properties:
        transactionId:
          type: string
        gatewayResponse:
          type: object
          additionalProperties: true
      required:
        - transactionId
    Withdrawal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        bankDetails:
          type: string
        upiId:
          type: string
        notes:
          type: string
        processedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - amount
        - status
        - createdAt
    WithdrawalRequest:
      type: object
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [bank_transfer, upi]
        bankDetails:
          type: string
        upiId:
          type: string
      required:
        - amount
        - method
...